Refactor petlist filter logic into petlist circuit
