+++ samples/star/src/main/kotlin/com/slack/circuit/star/home/HomePresenter.kt
@@ -15,49 +15,26 @@
  */
 package com.slack.circuit.star.home
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.systemBarsPadding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.selection.selectableGroup
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.FilterList
-import androidx.compose.material3.Button
-import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment.Companion.End
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitContent
 import com.slack.circuit.CircuitUiEvent
 import com.slack.circuit.CircuitUiState
-import com.slack.circuit.LocalOverlayHost
 import com.slack.circuit.NavEvent
 import com.slack.circuit.Navigator
-import com.slack.circuit.OverlayHost
 import com.slack.circuit.Presenter
 import com.slack.circuit.Screen
 import com.slack.circuit.ScreenUi
@@ -66,13 +43,6 @@ import com.slack.circuit.onNavEvent
 import com.slack.circuit.star.di.AppScope
 import com.slack.circuit.star.home.HomeScreen.Event.ChildNav
 import com.slack.circuit.star.home.HomeScreen.Event.HomeEvent
-import com.slack.circuit.star.home.HomeScreen.Event.PetListFilterEvent
-import com.slack.circuit.star.overlay.BottomSheetOverlay
-import com.slack.circuit.star.petlist.Filters
-import com.slack.circuit.star.petlist.Gender
-import com.slack.circuit.star.petlist.PetListFilterPresenter
-import com.slack.circuit.star.petlist.PetListFilterScreen
-import com.slack.circuit.star.petlist.Size
 import com.slack.circuit.star.ui.StarTheme
 import com.slack.circuit.ui
 import com.squareup.anvil.annotations.ContributesMultibinding
@@ -86,13 +56,11 @@ import kotlinx.parcelize.Parcelize
 object HomeScreen : Screen {
   data class State(
     val homeNavState: HomeNavScreen.State,
-    val petListFilterState: PetListFilterScreen.State,
     val eventSink: (Event) -> Unit,
   ) : CircuitUiState
 
   sealed interface Event : CircuitUiEvent {
     class HomeEvent(val event: HomeNavScreen.Event) : Event
-    class PetListFilterEvent(val event: PetListFilterScreen.Event) : Event
     class ChildNav(val navEvent: NavEvent) : Event
   }
 }
@@ -115,18 +83,14 @@ class HomePresenter
 @AssistedInject
 constructor(
   @Assisted private val navigator: Navigator,
-  petListFilterPresenterFactory: PetListFilterPresenter.Factory,
 ) : Presenter<HomeScreen.State> {
-  private val petListFilterPresenter = petListFilterPresenterFactory.create()
 
   @Composable
   override fun present(): HomeScreen.State {
     val homeNavState = HomeNavPresenter()
-    val petListFilterState = petListFilterPresenter.present()
-    return HomeScreen.State(homeNavState, petListFilterState) { event ->
+    return HomeScreen.State(homeNavState) { event ->
       when (event) {
         is HomeEvent -> homeNavState.eventSink(event.event)
-        is PetListFilterEvent -> petListFilterState.eventSink(event.event)
         is ChildNav -> navigator.onNavEvent(event.navEvent)
       }
     }
@@ -157,38 +121,8 @@ fun HomeContent(state: HomeScreen.State) {
   systemUiController.setNavigationBarColor(MaterialTheme.colorScheme.primaryContainer)
 
   val eventSink = state.eventSink
-  if (state.petListFilterState.showUpdateFiltersModal) {
-    val overlayHost = LocalOverlayHost.current
-    LaunchedEffect(state.petListFilterState.showUpdateFiltersModal) {
-      val result = overlayHost.updateFilters(state.petListFilterState)
-      eventSink(PetListFilterEvent(PetListFilterScreen.Event.UpdatedFilters(result)))
-    }
-  }
-
   Scaffold(
     modifier = Modifier.navigationBarsPadding().systemBarsPadding().fillMaxWidth(),
-    topBar = {
-      CenterAlignedTopAppBar(
-        title = {
-          Text("Adoptables", fontSize = 22.sp, color = MaterialTheme.colorScheme.onBackground)
-        },
-        colors =
-          TopAppBarDefaults.centerAlignedTopAppBarColors(
-            containerColor = MaterialTheme.colorScheme.background
-          ),
-        actions = {
-          IconButton(
-            onClick = { eventSink(PetListFilterEvent(PetListFilterScreen.Event.UpdateFilters)) }
-          ) {
-            Icon(
-              imageVector = Icons.Default.FilterList,
-              contentDescription = "Filter pet list",
-              tint = MaterialTheme.colorScheme.onBackground
-            )
-          }
-        },
-      )
-    },
     bottomBar = {
       StarTheme(useDarkTheme = true) {
         BottomNavigationBar(selectedIndex = state.homeNavState.index) { index ->
@@ -198,10 +132,7 @@ fun HomeContent(state: HomeScreen.State) {
     }
   ) { paddingValues ->
     Box(modifier = Modifier.padding(paddingValues)) {
-      val screen =
-        state.homeNavState.bottomNavItems[state.homeNavState.index].screenFor(
-          state.petListFilterState.filters
-        )
+      val screen = state.homeNavState.bottomNavItems[state.homeNavState.index].screen
       CircuitContent(screen, { event -> eventSink(ChildNav(event)) })
     }
   }
@@ -222,59 +153,6 @@ private fun BottomNavigationBar(selectedIndex: Int, onSelectedIndex: (Int) -> Un
         selected = selectedIndex == index,
         onClick = { onSelectedIndex(index) }
       )
-    }
-  }
-}
-
-private suspend fun OverlayHost.updateFilters(state: PetListFilterScreen.State): Filters {
-  return show(
-    BottomSheetOverlay(
-      model = state.filters,
-      onDismiss = { state.filters },
-    ) { initialFilters, overlayNavigator ->
-      var filters by remember { mutableStateOf(initialFilters) }
-      Column(Modifier.fillMaxWidth()) {
-        GenderFilterOption(filters.gender) { filters = filters.copy(gender = it) }
-        SizeFilterOption(filters.size) { filters = filters.copy(size = it) }
-
-        Row(Modifier.align(End)) {
-          Button(onClick = { overlayNavigator.finish(initialFilters) }) { Text("Cancel") }
-          Spacer(Modifier.width(16.dp))
-          Button(onClick = { overlayNavigator.finish(filters) }) { Text("Save") }
-        }
-      }
-    }
-  )
-}
-
-@Composable
-private fun GenderFilterOption(
-  selected: Gender,
-  selectedGender: (Gender) -> Unit,
-) {
-  Box { Text(text = "Gender") }
-  Row(modifier = Modifier.selectableGroup(), horizontalArrangement = Arrangement.SpaceEvenly) {
-    Gender.values().forEach { gender ->
-      Column {
-        Text(text = gender.name)
-        RadioButton(selected = selected == gender, onClick = { selectedGender(gender) })
-      }
-    }
-  }
-}
-
-@Composable
-private fun SizeFilterOption(
-  selected: Size,
-  selectedSize: (Size) -> Unit,
-) {
-  Box { Text(text = "Size") }
-  Row(modifier = Modifier.selectableGroup(), horizontalArrangement = Arrangement.SpaceEvenly) {
-    Size.values().forEach { size ->
-      Column {
-        Text(text = size.name)
-        RadioButton(selected = selected == size, onClick = { selectedSize(size) })
-      }
     }
   }
 }
