+++ samples/star/src/main/kotlin/com/slack/circuit/star/petlist/PetList.kt
@@ -22,24 +22,40 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.FilterList
+import androidx.compose.material3.Button
 import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ElevatedCard
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
@@ -50,12 +66,15 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.capitalize
 import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitUiEvent
 import com.slack.circuit.CircuitUiState
+import com.slack.circuit.LocalOverlayHost
 import com.slack.circuit.Navigator
+import com.slack.circuit.OverlayHost
 import com.slack.circuit.Presenter
 import com.slack.circuit.Screen
 import com.slack.circuit.ScreenUi
@@ -64,6 +83,7 @@ import com.slack.circuit.retained.produceRetainedState
 import com.slack.circuit.star.R
 import com.slack.circuit.star.data.Animal
 import com.slack.circuit.star.di.AppScope
+import com.slack.circuit.star.overlay.BottomSheetOverlay
 import com.slack.circuit.star.petdetail.PetDetailScreen
 import com.slack.circuit.star.petlist.PetListTestConstants.CARD_TAG
 import com.slack.circuit.star.petlist.PetListTestConstants.GRID_TAG
@@ -106,18 +126,22 @@ enum class Size {
 data class Filters(val gender: Gender = Gender.ALL, val size: Size = Size.ALL) : Parcelable
 
 @Parcelize
-data class PetListScreen(val filters: Filters = Filters()) : Screen {
+object PetListScreen : Screen {
   sealed interface State : CircuitUiState {
     object Loading : State
     object NoAnimals : State
     data class Success(
       val animals: List<PetListAnimal>,
+      val filters: Filters = Filters(),
+      val isUpdateFiltersModalShowing: Boolean = false,
       val eventSink: (Event) -> Unit,
     ) : State
   }
 
   sealed interface Event : CircuitUiEvent {
     data class ClickAnimal(val petId: Long, val photoUrlMemoryCacheKey: String?) : Event
+    object UpdateFilters : Event
+    data class UpdatedFilters(val newFilters: Filters) : Event
   }
 }
 
@@ -132,7 +156,7 @@ constructor(
     navigator: Navigator,
     circuitConfig: CircuitConfig
   ): Presenter<*>? {
-    if (screen is PetListScreen) return petListPresenterFactory.create(navigator, screen)
+    if (screen is PetListScreen) return petListPresenterFactory.create(navigator)
     return null
   }
 }
@@ -141,7 +165,6 @@ class PetListPresenter
 @AssistedInject
 constructor(
   @Assisted private val navigator: Navigator,
-  @Assisted private val screen: PetListScreen,
   private val petRepo: PetRepository,
 ) : Presenter<PetListScreen.State> {
   @Composable
@@ -152,24 +175,36 @@ constructor(
         value = animals.map { it.toPetListAnimal() }
       }
 
+    var isUpdateFiltersModalShowing by rememberSaveable { mutableStateOf(false) }
+    var filters by rememberSaveable { mutableStateOf(Filters()) }
+
     val animals = animalState
-    when {
+    return when {
       animals == null -> PetListScreen.State.Loading
       animals.isEmpty() -> PetListScreen.State.NoAnimals
       else ->
-        PetListScreen.State.Success(animals = animals.filter(::shouldKeep)) { event ->
+        PetListScreen.State.Success(
+          animals = animals.filter { shouldKeep(filters, it) },
+          filters = filters,
+          isUpdateFiltersModalShowing = isUpdateFiltersModalShowing,
+        ) { event ->
           when (event) {
             is PetListScreen.Event.ClickAnimal -> {
               navigator.goTo(PetDetailScreen(event.petId, event.photoUrlMemoryCacheKey))
+            }
+            is PetListScreen.Event.UpdatedFilters -> {
+              filters = event.newFilters
+            }
+            PetListScreen.Event.UpdateFilters -> {
+              isUpdateFiltersModalShowing = true
             }
           }
         }
     }
   }
 
-  private fun shouldKeep(animal: PetListAnimal): Boolean {
-    return screen.filters.gender.shouldKeep(animal.gender) &&
-      screen.filters.size.shouldKeep(animal.size)
+  private fun shouldKeep(filters: Filters, animal: PetListAnimal): Boolean {
+    return filters.gender.shouldKeep(animal.gender) && filters.size.shouldKeep(animal.size)
   }
 
   private fun Gender.shouldKeep(gender: String): Boolean {
@@ -184,7 +219,7 @@ constructor(
 
   @AssistedFactory
   interface Factory {
-    fun create(navigator: Navigator, screen: PetListScreen): PetListPresenter
+    fun create(navigator: Navigator): PetListPresenter
   }
 }
 
@@ -226,8 +261,40 @@ internal fun PetList(
   modifier: Modifier = Modifier,
   state: PetListScreen.State,
 ) {
+  if (state is PetListScreen.State.Success && state.isUpdateFiltersModalShowing) {
+    val eventSink = state.eventSink
+    val overlayHost = LocalOverlayHost.current
+    LaunchedEffect(state) {
+      val result = overlayHost.updateFilters(state.filters)
+      eventSink(PetListScreen.Event.UpdatedFilters(result))
+    }
+  }
+
   Scaffold(
     modifier = modifier,
+    topBar = {
+      CenterAlignedTopAppBar(
+        title = {
+          Text("Adoptables", fontSize = 22.sp, color = MaterialTheme.colorScheme.onBackground)
+        },
+        colors =
+          TopAppBarDefaults.centerAlignedTopAppBarColors(
+            containerColor = MaterialTheme.colorScheme.background
+          ),
+        actions = {
+          if (state is PetListScreen.State.Success) {
+            val eventSink = state.eventSink
+            IconButton(onClick = { eventSink(PetListScreen.Event.UpdateFilters) }) {
+              Icon(
+                imageVector = Icons.Default.FilterList,
+                contentDescription = "Filter pet list",
+                tint = MaterialTheme.colorScheme.onBackground
+              )
+            }
+          }
+        },
+      )
+    },
   ) { paddingValues ->
     when (state) {
       PetListScreen.State.Loading ->
@@ -315,6 +382,59 @@ private fun PetListGridItem(animal: PetListAnimal, onClick: () -> Unit) {
             style = MaterialTheme.typography.bodySmall,
           )
         }
+      }
+    }
+  }
+}
+
+private suspend fun OverlayHost.updateFilters(currentFilters: Filters): Filters {
+  return show(
+    BottomSheetOverlay(
+      model = currentFilters,
+      onDismiss = { currentFilters },
+    ) { initialFilters, overlayNavigator ->
+      var filters by remember { mutableStateOf(initialFilters) }
+      Column(Modifier.fillMaxWidth()) {
+        GenderFilterOption(filters.gender) { filters = filters.copy(gender = it) }
+        SizeFilterOption(filters.size) { filters = filters.copy(size = it) }
+
+        Row(Modifier.align(Alignment.End)) {
+          Button(onClick = { overlayNavigator.finish(initialFilters) }) { Text("Cancel") }
+          Spacer(Modifier.width(16.dp))
+          Button(onClick = { overlayNavigator.finish(filters) }) { Text("Save") }
+        }
+      }
+    }
+  )
+}
+
+@Composable
+private fun GenderFilterOption(
+  selected: Gender,
+  selectedGender: (Gender) -> Unit,
+) {
+  Box { Text(text = "Gender") }
+  Row(modifier = Modifier.selectableGroup(), horizontalArrangement = Arrangement.SpaceEvenly) {
+    Gender.values().forEach { gender ->
+      Column {
+        Text(text = gender.name)
+        RadioButton(selected = selected == gender, onClick = { selectedGender(gender) })
+      }
+    }
+  }
+}
+
+@Composable
+private fun SizeFilterOption(
+  selected: Size,
+  selectedSize: (Size) -> Unit,
+) {
+  Box { Text(text = "Size") }
+  Row(modifier = Modifier.selectableGroup(), horizontalArrangement = Arrangement.SpaceEvenly) {
+    Size.values().forEach { size ->
+      Column {
+        Text(text = size.name)
+        RadioButton(selected = selected == size, onClick = { selectedSize(size) })
       }
     }
   }
